generic batch inverse for slice types
