+++ src/core/fields/mod.rs
@@ -36,6 +36,29 @@ pub trait FieldExpOps: Mul<Output = Self> + MulAssign + Sized + One + Copy {
     }
 
     fn inverse(&self) -> Self;
+
+    // TODO(Ohad): Optimize (remove dependencies).
+    fn slice_batch_inverse(column: &[Self], dst: &mut [Self]) {
+        let n = column.len();
+        debug_assert!(n.is_power_of_two());
+        debug_assert!(dst.len() >= n);
+
+        dst[0] = column[0];
+        // First pass.
+        for i in 1..n {
+            dst[i] = dst[i - 1] * column[i];
+        }
+
+        // Inverse cumulative product.
+        let mut curr_inverse = dst[n - 1].inverse();
+
+        // Second pass.
+        for i in (1..n).rev() {
+            dst[i] = dst[i - 1] * curr_inverse;
+            curr_inverse *= column[i];
+        }
+        dst[0] = curr_inverse;
+    }
 }
 
 pub type Col<B, F> = <B as FieldOps<F>>::Column;
@@ -371,4 +394,27 @@ macro_rules! impl_extension_field {
             }
         }
     };
+}
+
+#[cfg(test)]
+mod tests {
+    use num_traits::Zero;
+    use rand::rngs::StdRng;
+    use rand::{Rng, SeedableRng};
+
+    use crate::core::fields::m31::{M31, P};
+    use crate::core::fields::FieldExpOps;
+
+    #[test]
+    fn test_slice_batch_inverse() {
+        let mut rng = StdRng::seed_from_u64(0);
+        let elements: Vec<M31> = (0..16)
+            .map(|_| M31::from_u32_unchecked(rng.gen::<u32>() % P))
+            .collect();
+        let expected = elements.iter().map(|e| e.inverse()).collect::<Vec<_>>();
+        let mut dst = [M31::zero(); 16];
+
+        M31::slice_batch_inverse(&elements, &mut dst);
+        assert_eq!(expected, dst);
+    }
 }
